refactor(analytics): extract analytics-aggregation types and drop dead viral-hooks method
analytics-aggregation.service.ts exceeded the 1000-line runtime-complexity
guard. Extract the 12 module-scope interfaces/types into a sibling
analytics-aggregation.types.ts (re-exported for compatibility), and delete
getViralHooksSummary along with its exclusively-used helpers
(buildAliasedBrandSqlPredicate, PostViewsRow import) — the method had zero
callers and zero test coverage; the live viral-hooks feature is implemented
independently in endpoints/analytics/analytics.service.ts. File drops from
1262 to 909 lines; getPlatformAnalytics (used by two controllers) is
untouched.
